Move Components::Sidebar::* to Views::Layouts::Sidebar::*
The five sidebar components (Section, Admin, Login, User, Languages)
only render from `Views::Layouts::ApplicationSidebar` and live in
the application layout — not reusable building blocks, layout-bound
content. Files move from `app/components/sidebar/` to
`app/views/layouts/sidebar/`, namespace `Views::Layouts::Sidebar::*`
to match the `Views::Layouts` umbrella introduced in #4362
(cf. existing `Views::Layouts::ApplicationSidebar` and
`Views::Layouts::LoginWelcome`).

Tests move alongside: `test/components/sidebar/` →
`test/views/layouts/sidebar/`. Each test's `module Sidebar` becomes
`module Views::Layouts::Sidebar`, and `Components::Sidebar::Foo.new`
references inside become unqualified `Foo.new` (lexical sibling
resolution within the module).

The sole external caller — `Views::Layouts::ApplicationSidebar` —
updates all nine renders to `Views::Layouts::Sidebar::*`. Because
`ApplicationSidebar` uses one-line class declaration
(`class Views::Layouts::ApplicationSidebar < Views::Base`) rather
than a nested `module Views::Layouts; class ApplicationSidebar` body,
its lexical scope doesn't include `Views::Layouts` — unqualified
`Sidebar::*` wouldn't resolve. Full path is used instead, same
verbosity as the old `Components::Sidebar::*`.
