+++ test/views/layouts/sidebar/indexes_test.rb
@@ -2,9 +2,9 @@
 
 require "test_helper"
 
-module Sidebar
-  class SpeciesListsTest < ComponentTestCase
-    include Tabs::Sidebar::SpeciesListsHelper
+module Views::Layouts::Sidebar
+  class IndexesTest < ComponentTestCase
+    include Tabs::Sidebar::IndexesHelper
     include Rails.application.routes.url_helpers
 
     def setup
@@ -24,13 +24,15 @@ module Sidebar
     def test_renders_heading_and_links
       html = render_component
 
-      # Should have heading with "Species Lists:" text
-      assert_includes(html, :app_species_list.t)
+      # Should have heading with "Indexes:" text
+      assert_includes(html, :INDEXES.t)
 
       # Should include navigation links
-      assert_html(html, "#nav_your_species_lists_link")
-      assert_html(html, "#nav_species_lists_link")
-      assert_html(html, "#nav_new_species_list_link")
+      assert_html(html, "#nav_articles_link")
+      assert_html(html, "#nav_herbaria_link")
+      assert_html(html, "#nav_locations_link")
+      assert_html(html, "#nav_name_observations_link")
+      assert_html(html, "#nav_projects_link")
 
       # Should have indent class on links
       assert_html(html, ".list-group-item.indent")
@@ -46,30 +48,16 @@ module Sidebar
       assert_html(html, ".list-group-item.disabled.font-weight-bold")
     end
 
-    def test_renders_only_all_lists_link_for_guest_users
-      html = render_component(user: nil)
-
-      # Should have heading
-      assert_includes(html, :app_species_list.t)
-
-      # Should have all lists link (available to all)
-      assert_html(html, "#nav_species_lists_link")
-
-      # Should NOT have user-only links
-      assert_no_html(html, "#nav_your_species_lists_link")
-      assert_no_html(html, "#nav_new_species_list_link")
-    end
-
     private
 
-    def render_component(user: users(:rolf))
+    def render_component
       classes = {
         heading: "list-group-item disabled font-weight-bold",
         indent: "list-group-item indent"
       }
-      render(Components::Sidebar::Section.new(
-               heading_key: :app_species_list,
-               tabs: sidebar_species_lists_tabs(user),
+      render(Section.new(
+               heading_key: :INDEXES,
+               tabs: sidebar_indexes_tabs,
                classes: classes
              ))
     end
