+++ test/views/layouts/sidebar/latest_test.rb
@@ -2,9 +2,9 @@
 
 require "test_helper"
 
-module Sidebar
-  class IndexesTest < ComponentTestCase
-    include Tabs::Sidebar::IndexesHelper
+module Views::Layouts::Sidebar
+  class LatestTest < ComponentTestCase
+    include Tabs::Sidebar::LatestHelper
     include Rails.application.routes.url_helpers
 
     def setup
@@ -24,15 +24,14 @@ module Sidebar
     def test_renders_heading_and_links
       html = render_component
 
-      # Should have heading with "Indexes:" text
-      assert_includes(html, :INDEXES.t)
+      # Should have heading with "Latest:" text
+      assert_includes(html, :app_latest.t)
 
       # Should include navigation links
       assert_html(html, "#nav_articles_link")
-      assert_html(html, "#nav_herbaria_link")
-      assert_html(html, "#nav_locations_link")
-      assert_html(html, "#nav_name_observations_link")
-      assert_html(html, "#nav_projects_link")
+      assert_html(html, "#nav_activity_logs_link")
+      assert_html(html, "#nav_images_link")
+      assert_html(html, "#nav_comments_link")
 
       # Should have indent class on links
       assert_html(html, ".list-group-item.indent")
@@ -48,16 +47,31 @@ module Sidebar
       assert_html(html, ".list-group-item.disabled.font-weight-bold")
     end
 
+    def test_renders_only_news_link_for_guest_users
+      html = render_component(user: nil)
+
+      # Should have heading
+      assert_includes(html, :app_latest.t)
+
+      # Should have news link (available to all)
+      assert_html(html, "#nav_articles_link")
+
+      # Should NOT have user-only links
+      assert_no_html(html, "#nav_activity_logs_link")
+      assert_no_html(html, "#nav_images_link")
+      assert_no_html(html, "#nav_comments_link")
+    end
+
     private
 
-    def render_component
+    def render_component(user: users(:rolf))
       classes = {
         heading: "list-group-item disabled font-weight-bold",
         indent: "list-group-item indent"
       }
-      render(Components::Sidebar::Section.new(
-               heading_key: :INDEXES,
-               tabs: sidebar_indexes_tabs,
+      render(Section.new(
+               heading_key: :app_latest,
+               tabs: sidebar_latest_tabs(user),
                classes: classes
              ))
     end
