+++ test/views/layouts/sidebar/species_lists_test.rb
@@ -2,9 +2,9 @@
 
 require "test_helper"
 
-module Sidebar
-  class ObservationsTest < ComponentTestCase
-    include Tabs::Sidebar::ObservationsHelper
+module Views::Layouts::Sidebar
+  class SpeciesListsTest < ComponentTestCase
+    include Tabs::Sidebar::SpeciesListsHelper
     include Rails.application.routes.url_helpers
 
     def setup
@@ -24,14 +24,13 @@ module Sidebar
     def test_renders_heading_and_links
       html = render_component
 
-      # Should have heading with "Observations:" text
-      assert_includes(html, :app_observations_left.t)
+      # Should have heading with "Species Lists:" text
+      assert_includes(html, :app_species_list.t)
 
       # Should include navigation links
-      assert_html(html, "#nav_observations_link")
-      assert_html(html, "#nav_new_observation_link")
-      assert_html(html, "#nav_your_observations_link")
-      assert_html(html, "#nav_identify_observations_link")
+      assert_html(html, "#nav_your_species_lists_link")
+      assert_html(html, "#nav_species_lists_link")
+      assert_html(html, "#nav_new_species_list_link")
 
       # Should have indent class on links
       assert_html(html, ".list-group-item.indent")
@@ -47,19 +46,18 @@ module Sidebar
       assert_html(html, ".list-group-item.disabled.font-weight-bold")
     end
 
-    def test_renders_only_latest_link_for_guest_users
+    def test_renders_only_all_lists_link_for_guest_users
       html = render_component(user: nil)
 
       # Should have heading
-      assert_includes(html, :app_observations_left.t)
+      assert_includes(html, :app_species_list.t)
 
-      # Should have latest observations link (available to all)
-      assert_html(html, "#nav_observations_link")
+      # Should have all lists link (available to all)
+      assert_html(html, "#nav_species_lists_link")
 
       # Should NOT have user-only links
-      assert_no_html(html, "#nav_new_observation_link")
-      assert_no_html(html, "#nav_your_observations_link")
-      assert_no_html(html, "#nav_identify_observations_link")
+      assert_no_html(html, "#nav_your_species_lists_link")
+      assert_no_html(html, "#nav_new_species_list_link")
     end
 
     private
@@ -69,9 +67,9 @@ module Sidebar
         heading: "list-group-item disabled font-weight-bold",
         indent: "list-group-item indent"
       }
-      render(Components::Sidebar::Section.new(
-               heading_key: :app_observations_left,
-               tabs: sidebar_observations_tabs(user),
+      render(Section.new(
+               heading_key: :app_species_list,
+               tabs: sidebar_species_lists_tabs(user),
                classes: classes
              ))
     end
